test(LottoFixture.java) : @MothodSource Fixture로 이동
