+++ src/test/java/lotto/domain/LottoTest.java
@@ -2,13 +2,9 @@ package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.Arrays;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.stream.Stream;
 
 import static lotto.domain.Lotto.DUPLICATE_EXCEPTION_MESSAGE;
 import static lotto.domain.Lotto.LOTTO_SIZE;
@@ -20,14 +16,14 @@ class LottoTest {
 
     @DisplayName("로또 생성")
     @ParameterizedTest
-    @MethodSource("constructor")
+    @MethodSource("lotto.fixture.LottoFixture#constructor")
     void constructor(List<Integer> lottoNumbers) {
         assertThatNoException().isThrownBy(() -> new Lotto(lottoNumbers));
     }
 
     @DisplayName("6개 이상의 수를 추가 할 수 없다.")
     @ParameterizedTest
-    @MethodSource("lottoSize")
+    @MethodSource("lotto.fixture.LottoFixture#lottoSize")
     void maxSize(List<Integer> lottoNumbers) {
         assertThatThrownBy(() -> new Lotto(lottoNumbers))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -36,29 +32,10 @@ class LottoTest {
 
     @DisplayName("중복된 수를 추가할 수 없다.")
     @ParameterizedTest
-    @MethodSource("duplicate")
+    @MethodSource("lotto.fixture.LottoFixture#duplicate")
     void duplicate(List<Integer> lottoNumbers) {
         assertThatThrownBy(() -> new Lotto(lottoNumbers))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(DUPLICATE_EXCEPTION_MESSAGE);
-    }
-
-    private static Stream<Arguments> lottoSize() {
-        List<Arguments> listOfArguments = new LinkedList<>();
-        listOfArguments.add(Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6, 7)));
-        listOfArguments.add(Arguments.of(Arrays.asList(1, 2, 3, 4, 5)));
-        return listOfArguments.stream();
-    }
-
-    private static Stream<Arguments> duplicate() {
-        List<Arguments> listOfArguments = new LinkedList<>();
-        listOfArguments.add(Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 5)));
-        return listOfArguments.stream();
-    }
-
-    private static Stream<Arguments> constructor() {
-        List<Arguments> listOfArguments = new LinkedList<>();
-        listOfArguments.add(Arguments.of(Arrays.asList(1, 2, 3, 4, 5, 6)));
-        return listOfArguments.stream();
     }
 }
