+++ src/test/java/lotto/domain/WinningNumberTest.java
@@ -1,12 +1,10 @@
 package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.stream.Stream;
 
 import static lotto.domain.Lotto.DUPLICATE_EXCEPTION_MESSAGE;
 import static lotto.domain.WinningNumber.WINNING_NUMBER;
@@ -15,6 +13,7 @@ import static lotto.fixture.LottoFixture.로또번호123457;
 import static lotto.fixture.WinningNumberFixture.당첨번호123456;
 import static org.assertj.core.api.Assertions.*;
 
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @DisplayName("당첨번호")
 class WinningNumberTest {
 
@@ -34,7 +33,7 @@ class WinningNumberTest {
 
     @DisplayName("6자리 미만일 수 없다.")
     @ParameterizedTest
-    @MethodSource("size")
+    @MethodSource("lotto.fixture.WinningNumberFixture#size")
     void size(String[] numbers) {
         assertThatThrownBy(() -> new WinningNumber(numbers))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -43,7 +42,7 @@ class WinningNumberTest {
 
     @DisplayName("중복된 수를 입력할 수 없다.")
     @ParameterizedTest
-    @MethodSource("duplicate")
+    @MethodSource("lotto.fixture.WinningNumberFixture#duplicate")
     void duplicateNumber(String[] numbers) {
         assertThatThrownBy(() -> new WinningNumber(numbers))
                 .isInstanceOf(IllegalArgumentException.class)
@@ -52,7 +51,7 @@ class WinningNumberTest {
 
     @DisplayName("숫자만 입력 가능하다.")
     @ParameterizedTest
-    @MethodSource("type")
+    @MethodSource("lotto.fixture.WinningNumberFixture#type")
     void type(String[] numbers) {
         assertThatThrownBy(() -> new WinningNumber(numbers))
                 .isInstanceOf(NumberFormatException.class);
@@ -60,34 +59,8 @@ class WinningNumberTest {
 
     @DisplayName("숫자만 입력 가능하다.")
     @ParameterizedTest
-    @MethodSource("constructor")
+    @MethodSource("lotto.fixture.WinningNumberFixture#constructor")
     void constructor(String[] numbers) {
         assertThatNoException().isThrownBy(() -> new WinningNumber(numbers));
-    }
-
-    static Stream<Arguments> size() {
-        return Stream.of(
-                Arguments.of((Object) new String[]{"1", "2", "3", "4"}),
-                Arguments.of((Object) new String[]{"1", "2", "3", "4", "5"}),
-                Arguments.of((Object) new String[]{"1", "2", "3", "4", "5", "6", "7"})
-        );
-    }
-
-    static Stream<Arguments> duplicate() {
-        return Stream.of(
-                Arguments.of((Object) new String[]{"1", "2", "3", "4", "5", "5"})
-        );
-    }
-
-    static Stream<Arguments> type() {
-        return Stream.of(
-                Arguments.of((Object) new String[]{"1", "2", "3", "4", "5", "x"})
-        );
-    }
-
-    static Stream<Arguments> constructor() {
-        return Stream.of(
-                Arguments.of((Object) new String[]{"1", "2", "3", "4", "5", "6"})
-        );
     }
 }
